Mcr#63

+++ src/main/java/ru/isaykin/controllers/AuthorController.java
@@ -19,14 +19,14 @@ public class AuthorController {
     public AuthorController(AuthorsSQLService authorsSQLService) {
         this.authorsSQLService = authorsSQLService;
     }
-
+//TODO: ADD SPRING VALIDATION
 
     @GetMapping("authors")
     public ResponseEntity<Object> getListOrGetOneByFirstNameAndLastName(@RequestParam(value = "first_name", required = false) String firstName,
                                                                         @RequestParam(value = "last_name", required = false) String lastName) {
         ResponseEntity<Object> result;
         if (firstName != null || lastName != null) {
-            List<Author> resultList = authorsSQLService.getListByFirstNameAndLastName(firstName, lastName);
+            List<Author> resultList = authorsSQLService.getListByFirstNameAndLastNameOrNull(firstName, lastName);
             if (resultList != null) {
                 result = new ResponseEntity<>(resultList, OK);
             } else {
@@ -41,29 +41,49 @@ public class AuthorController {
 
     @GetMapping("authors/{id}")
     public ResponseEntity<Author> getOneAuthor(@PathVariable Long id) {
-        return authorsSQLService.getOneById(id);
+        ResponseEntity<Author> responseEntity;
+        if (id != null) {
+            responseEntity = authorsSQLService.getOneById(id);
+        } else {
+            responseEntity = new ResponseEntity<>(NOT_FOUND);
+        }
+        return responseEntity;
     }
 
 
     @GetMapping("authors/age/gt/{age}")
-    public ResponseEntity<List<Author>> getListByAgeGraterThen(@PathVariable int age) {
-        List<Author> authorList = authorsSQLService.getListByAgeGT(age);
-        if (authorList.isEmpty()) {
-            return new ResponseEntity<>(NOT_FOUND);
-        } else {
-            return new ResponseEntity<>(authorList, OK);
+    public ResponseEntity<List<Author>> getListByAgeGraterThen(@PathVariable Integer age) {
+        ResponseEntity<List<Author>> responseEntity;
+        if (age != null) {
+            List<Author> authorList = authorsSQLService.getListByAgeGT(age);
+            if (authorList.isEmpty()) {
+                responseEntity = new ResponseEntity<>(NOT_FOUND);
+            } else {
+                responseEntity = new ResponseEntity<>(authorList, OK);
+            }
+        }else {
+            responseEntity = new ResponseEntity<>(NOT_FOUND);
         }
+        return responseEntity;
     }
 
     @GetMapping("authors/age/lt/{age}")
-    public ResponseEntity<List<Author>> getListByAgeLessThen(@PathVariable int age) {
-        List<Author> authorList = authorsSQLService.getListByAgeLT(age);
-        if (authorList.isEmpty()) {
-            return new ResponseEntity<>(NOT_FOUND);
+    public ResponseEntity<List<Author>> getListByAgeLessThen(@PathVariable Integer age) {
+        ResponseEntity<List<Author>> responseEntity;
+        if (age != null) {
+            List<Author> authorList = authorsSQLService.getListByAgeLT(age);
+
+            if (authorList.isEmpty()) {
+                responseEntity = new ResponseEntity<>(NOT_FOUND);
+            } else {
+                responseEntity = new ResponseEntity<>(authorList, OK);
+            }
         } else {
-            return new ResponseEntity<>(authorList, OK);
+            responseEntity = new ResponseEntity<>(NOT_FOUND);
         }
+            return responseEntity;
     }
+
 
     @PostMapping("authors")
     public ResponseEntity<Author> insert(@RequestBody Author author) {
@@ -75,28 +95,32 @@ public class AuthorController {
             Author author1 = authorsSQLService.insert(author);
             result = new ResponseEntity<>(author1, OK);
         }
-
         return result;
     }
 
     @PostMapping("authors/addlist")
     public ResponseEntity<List<Author>> createList(@RequestBody List<Author> authorList) {
         List<Author> insertedList;
+        ResponseEntity<List<Author>> responseEntity;
         if (authorList == null) {
-            return new ResponseEntity<>(NO_CONTENT);
+            responseEntity = new ResponseEntity<>(NO_CONTENT);
         } else {
             insertedList = authorsSQLService.insertMany(authorList);
-            return new ResponseEntity<>(insertedList, OK);
+            responseEntity = new ResponseEntity<>(insertedList, OK);
         }
+        return responseEntity;
     }
 
 
     @DeleteMapping("authors/{id}")
     public ResponseEntity<Author> delete(@PathVariable Long id) {
-
-
-        return authorsSQLService.delete(id);
-
+        ResponseEntity<Author> responseEntity;
+        if (id != null) {
+            responseEntity = authorsSQLService.delete(id);
+        } else {
+            responseEntity = new ResponseEntity<>(NOT_FOUND);
+        }
+        return responseEntity;
     }
 
     @PutMapping("authors/{id}")
